Create mixin, migration
[#OSF-8275]
